requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
langchain==0.0.190
openai==0.27.0
pydantic==1.10.9

